+++ pkg/signatures/teddsa/frost/signing/aggregation/aggregation.go
@@ -130,7 +130,7 @@ func (sa *SignatureAggregator) Aggregate(partialSignatures map[integration.Ident
 
 	// identifiable abort is possible
 	if sa.HasIdentifiableAbort() {
-		shamirConfig, err := sharing.NewShamir(uint32(sa.CohortConfig.Threshold), uint32(sa.CohortConfig.TotalParties), sa.CohortConfig.CipherSuite.Curve)
+		shamirConfig, err := sharing.NewShamir(sa.CohortConfig.Threshold, sa.CohortConfig.TotalParties, sa.CohortConfig.CipherSuite.Curve)
 		if err != nil {
 			return nil, errors.Wrap(err, "could not initialize shamir config")
 		}
